Move Xxxable classes into the behavior directory
