+++ lib/src/behaviors/introduction_finish_loggable.dart
