+++ lib/src/blocs/card_create_bloc.dart
@@ -1,10 +1,10 @@
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 import '../entities/authentication_session.dart';
-import './card_creatable.dart';
-import './card_create_loggable.dart';
+import '../behaviors/card_creatable.dart';
+import '../behaviors/card_create_loggable.dart';
 export '../entities/authentication_session.dart';
-export './card_creatable.dart' show NewCardText;
+export '../behaviors/card_creatable.dart' show NewCardText;
 
 abstract class CardCreateBloc {
   ValueObservable<NewCardText> get text;
