+++ lib/src/blocs/card_list_bloc.dart
@@ -2,7 +2,7 @@ import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 import '../entities/authentication_session.dart';
 import '../entities/card.dart';
-import './card_subscribable.dart';
+import '../behaviors/card_subscribable.dart';
 export '../entities/card.dart';
 
 abstract class CardListBloc {
