+++ lib/src/blocs/introduction_bloc.dart
@@ -2,10 +2,10 @@ import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:saraka/constants.dart';
 import '../entities/card.dart';
-import './card_subscribable.dart';
-import './introduction_finish_loggable.dart';
-import './introduction_finishable.dart';
-import './introduction_page_change_loggable.dart';
+import '../behaviors/card_subscribable.dart';
+import '../behaviors/introduction_finish_loggable.dart';
+import '../behaviors/introduction_finishable.dart';
+import '../behaviors/introduction_page_change_loggable.dart';
 export '../entities/card.dart';
 
 abstract class IntroductionBloc {
