+++ lib/src/blocs/synthesizer_bloc.dart
@@ -1,10 +1,10 @@
 import 'dart:io' show File;
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
-import './sound_file_playable.dart';
-import './synthesizable.dart';
-import './synthesized_sound_file_referable.dart';
-import './synthesize_loggable.dart';
+import '../behaviors/sound_file_playable.dart';
+import '../behaviors/synthesizable.dart';
+import '../behaviors/synthesized_sound_file_referable.dart';
+import '../behaviors/synthesize_loggable.dart';
 
 abstract class SynthesizerBloc {
   Observable<bool> get isCaching;
