+++ lib/src/implementations/firestore_card_repository.dart
@@ -1,10 +1,10 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
-import '../blocs/card_deletable.dart';
-import '../blocs/card_subscribable.dart';
-import '../blocs/in_queue_card_subscribable.dart';
-import '../blocs/review_subscribable.dart';
+import '../behaviors/card_deletable.dart';
+import '../behaviors/card_subscribable.dart';
+import '../behaviors/in_queue_card_subscribable.dart';
+import '../behaviors/review_subscribable.dart';
 import './firestore_card.dart';
 import './firestore_review.dart';
